Write a program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,18 +7,18 @@
  */
 int main(void)
 {
-unsigned long long int fib1 = 1, fib2 = 2, nextFib;
+unsigned long int fib1 = 1, fib2 = 2, nextFib;
 int count, numPerLine;
 
 numPerLine = 10;
 
-printf("%llu, %llu", fib1, fib2);
+printf("%lu, %lu", fib1, fib2);
 
 for (count = 3; count <= 98; count++)
 {
 nextFib = fib1 + fib2;
 
-printf(", %llu", nextFib);
+printf(", %lu", nextFib);
 
 if (count % numPerLine == 0)
 printf("\n");
